% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_read_sodium.R
\name{cache_read_sodium}
\alias{cache_read_sodium}
\title{Read cache using sodium}
\usage{
cache_read_sodium(name, cache = cache_find(),
  key = getOption("cache.sodium_key", set_option(cache.sodium_key =
  readline("sodium Encryption Key? "))))
}
\arguments{
\item{name}{string; name of object in the cache}

\item{cache}{string; path to cache directory. Defaults to \code{\link[=cache_find]{cache_find()}}}

\item{key}{string; key for writing and resotring}
}
\description{
Read cache using sodium
}
\details{
\code{cache_read_sodium} reads an AES encrypted object from the cache. \code{AES}
encrypted objects are
}
