% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_register_sodium.R
\name{cache_register_sodium}
\alias{cache_register_sodium}
\title{Register sodium as a backend for cache}
\usage{
cache_register_sodium()
}
\description{
Installs \code{cache.sodium} for use as an extension for writing/reading files
}
\details{
You probably should use \code{\link[=cache_use_sodium]{cache_use_sodium()}} instead which makes sodium
the default method for reading/writing to the cache.  There is generally no
reason to invoke this function as it is called when the packages is
attached.
}
\seealso{
\itemize{
\item \code{\link[=cache_use_sodium]{cache_use_sodium()}}
}
}
