% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sodium_keys.R
\name{sodium_set_key}
\alias{sodium_set_key}
\alias{sodium_get_key}
\alias{sodium_get_key_or_ask}
\alias{cache_sodium_key}
\title{sodium key management}
\usage{
sodium_set_key(key)

sodium_get_key()

sodium_get_key_or_ask()

cache_sodium_key(key = NULL)
}
\arguments{
\item{key}{string for encryption key}
}
\description{
Get and set the key for encryption and decryption
}
\details{
\code{\link[=sodium_set_key]{sodium_set_key()}} sets the encryption key to \code{key}. It is stored as an option:
'cache.sodium.key'

\code{\link[=sodium_get_key]{sodium_get_key()}} retrieves the key. It returns NULL if no key has been set.

\code{\link[=sodium_get_key_or_ask]{sodium_get_key_or_ask()}} retrieves the key. If no key has been set and the
session is interactive, the user is prompted for the key.

\code{\link[=cache_sodium_key]{cache_sodium_key()}} can be used to set the sodium key if \code{key} is provided.
Otherwise it is retrieved.  This matches the \emph{cache-style} accessors.
}
\examples{

  sodium_set_key('xyzzy')
  sodium_get_key()
  sodium_set_key(NULL)
  sodium_get_key()

\dontrun{
  sodium_get_key_or_ask()
  sodium_get_key()
}

}
