% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_use_sodium.R
\name{cache_use_sodium}
\alias{cache_use_sodium}
\title{Cache objects using sodium (Sodium)}
\usage{
cache_use_sodium(key = NULL)
}
\arguments{
\item{key}{string; sets the en/decryption key \code{cache.sodium.key}.
The default (NULL) does not set a key.}
}
\description{
Caches objects with sodium encryption using the \emph{sodium} package
}
\details{
Encrypts cached datasets using sodium256 encryption provided by the \emph{sodium}
package. Cache

If \code{key} is provided, this is used as the encryption/decrytion key.
}
\examples{

dontrun{
 cache_use_sodium()
 cache(mtcars)
 if( exists('mtcars') ) rm(mtcars)
 uncache(mtcars)
 fs::dir_ls( cache_find() )
}

}
