% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sodium.R
\name{read_sodium}
\alias{read_sodium}
\title{Read cache using sodium}
\usage{
read_sodium(path, key = getOption("cache.sodium.key",
  set_option(cache.sodium.key = readline("sodium Encryption Key? "))))
}
\arguments{
\item{key}{string; key for writing and resotring}

\item{name}{string; name of object in the cache}

\item{cache}{string; path to cache directory. Defaults to \code{\link[=cache_find]{cache_find()}}}
}
\description{
Read cache using sodium
}
\details{
\code{cache_read_sodium} reads an AES encrypted object from the cache. \code{AES}
encrypted objects are
}
